requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
aiohttp==3.9.3
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
arrow==1.3.0
    # via -r requirements/production.in
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via html-sanitizer
billiard==4.2.0
    # via celery
boto3==1.34.52
    # via django-storages
botocore==1.34.52
    # via
    #   boto3
    #   django-s3direct
    #   s3transfer
celery[redis]==5.3.6
    # via
    #   -r requirements/production.in
    #   django-celery-beat
    #   django-import-export-extensions
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.5
    # via django-rest-knox
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
django==4.2.10
    # via
    #   -r requirements/production.in
    #   django-appconf
    #   django-autocomplete-light
    #   django-celery-beat
    #   django-citext
    #   django-ckeditor
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-import-export
    #   django-import-export-extensions
    #   django-js-asset
    #   django-picklefield
    #   django-redis
    #   django-rest-knox
    #   django-s3direct
    #   django-safedelete
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   drf-standardized-errors
    #   sentry-sdk
django-advanced-password-validation==1.2.0
    # via -r requirements/production.in
django-appconf==1.0.6
    # via django-imagekit
django-autocomplete-light==3.11.0
    # via -r requirements/production.in
django-celery-beat==2.5.0
    # via -r requirements/production.in
django-citext==1.0.0
    # via -r requirements/production.in
django-ckeditor==6.7.1
    # via -r requirements/production.in
django-extensions==3.2.3
    # via
    #   -r requirements/production.in
    #   django-import-export-extensions
django-filter==23.5
    # via
    #   -r requirements/production.in
    #   django-import-export-extensions
django-health-check==3.18.1
    # via -r requirements/production.in
django-imagekit==5.0.0
    # via -r requirements/production.in
django-import-export==3.3.7
    # via django-import-export-extensions
django-import-export-extensions==0.5.0
    # via -r requirements/production.in
django-js-asset==2.2.0
    # via django-ckeditor
django-ordered-model==3.7.4
    # via -r requirements/production.in
django-picklefield==3.1
    # via django-import-export-extensions
django-redis==5.4.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-s3direct @ git+https://github.com/Saritasa/django-s3direct.git
    # via -r requirements/production.in
django-safedelete==1.3.3
    # via -r requirements/production.in
django-storages[boto3]==1.14.2
    # via -r requirements/production.in
django-timezone-field==6.1.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/production.in
    #   django-import-export-extensions
    #   django-rest-knox
    #   django-s3direct
    #   drf-spectacular
    #   drf-standardized-errors
drf-spectacular==0.26.5
    # via
    #   -r requirements/production.in
    #   django-import-export-extensions
    #   drf-standardized-errors
drf-standardized-errors[openapi]==0.12.6
    # via -r requirements/production.in
et-xmlfile==1.1.0
    # via openpyxl
execnet==2.0.2
    # via pytest-xdist
factory-boy==3.3.0
    # via -r requirements/production.in
faker==23.3.0
    # via factory-boy
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
html-sanitizer==2.3.0
    # via -r requirements/production.in
idna==3.6
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-standardized-errors
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.5
    # via celery
lxml==5.1.0
    # via html-sanitizer
markdown-it-py==3.0.0
    # via rich
markuppy==1.14
    # via tablib
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via -r requirements/production.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mysqlclient==2.2.4
    # via -r requirements/production.in
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.2
    # via
    #   pytest
    #   pytest-sugar
pilkit==3.0
    # via django-imagekit
pillow==10.2.0
    # via
    #   -r requirements/production.in
    #   pilkit
pluggy==1.4.0
    # via pytest
prompt-toolkit==3.0.43
    # via click-repl
psycopg[binary]==3.1.18
    # via
    #   -r requirements/production.in
    #   django-citext
psycopg-binary==3.1.18
    # via psycopg
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via rich
pyjwt==2.8.0
    # via twilio
pytest==8.0.2
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-lazy-fixtures
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.8.0
    # via -r requirements/production.in
pytest-lazy-fixtures==1.0.5
    # via -r requirements/production.in
pytest-sugar==1.0.0
    # via -r requirements/production.in
pytest-xdist==3.5.0
    # via -r requirements/production.in
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   celery
    #   faker
    #   python-crontab
python-decouple==3.8
    # via -r requirements/production.in
pytz==2024.1
    # via djangorestframework
pyuwsgi==2.0.23.post0
    # via -r requirements/production.in
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   tablib
redis==5.0.2
    # via
    #   celery
    #   django-redis
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-import-export-extensions
    #   twilio
rich==13.7.1
    # via -r requirements/production.in
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.0
    # via boto3
sentry-sdk[django]==1.40.6
    # via -r requirements/production.in
six==1.16.0
    # via python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
termcolor==2.4.0
    # via pytest-sugar
twilio==9.0.0
    # via -r requirements/production.in
types-python-dateutil==2.8.19.20240106
    # via arrow
typing-extensions==4.10.0
    # via psycopg
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.7
    # via
    #   botocore
    #   requests
    #   sentry-sdk
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
yarl==1.9.4
    # via aiohttp
